ha_autogen/requirements.txt
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
aiohttp>=3.10.0,<4.0
httpx>=0.27.0,<1.0
ruamel.yaml>=0.18.0,<1.0
pydantic>=2.9.0,<3.0
aiosqlite>=0.20.0,<1.0
jinja2>=3.1.0,<4.0
python-multipart>=0.0.9

# Dev / test
pytest>=8.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.35.0
